+++ Firmware/communication/ascii_protocol.hpp
@@ -21,7 +21,7 @@ void ASCII_protocol_parse_stream(const uint8_t* buffer, size_t len, StreamSink& 
 #define NUM_ASCII_USER_VALUES 2
 
 struct UserValue {
-    float val = 0;
+    uint32_t val = 0;
     uint32_t timestamp = 0;
 };
